fix(RVCV-161): domain의 enum 타입을 value 패키지로 옮긴다
