+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/request/UpdateContainerRequest.java
@@ -2,8 +2,7 @@ package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.
 
 import com.romanticpipe.reviewcanvas.reviewproperty.domain.Border;
 import com.romanticpipe.reviewcanvas.reviewproperty.domain.Padding;
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.Shadow;
-
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.value.Shadow;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
