+++ api-module/src/main/java/com/romanticpipe/reviewcanvas/domain/reviewproperty/application/usecase/response/GetReviewLayoutResponse.java
@@ -1,9 +1,9 @@
 package com.romanticpipe.reviewcanvas.domain.reviewproperty.application.usecase.response;
 
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.FocusAreaLayout;
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.ImageReviewAreaLayout;
 import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewLayout;
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewLayoutDesign;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.value.FocusAreaLayout;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.value.ImageReviewAreaLayout;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.value.ReviewLayoutDesign;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record GetReviewLayoutResponse(@Schema(description = "베스트 리뷰 영역", requiredMode = Schema.RequiredMode.REQUIRED)
