--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/ButtonType.java
@@ -1,5 +1,0 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
-
-public enum ButtonType {
-	NONE, THUMB_UP_WITH_TEXT, THUMB_UP
-}
