--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/DetailViewType.java
@@ -1,5 +1,0 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
-
-public enum DetailViewType {
-	SPREAD, MODAL
-}
