--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/PagingType.java
@@ -1,5 +1,0 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
-
-public enum PagingType {
-	PAGE_NUMBER, SEE_MORE_SCROLL
-}
