--- domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/ReviewTitleType.java
@@ -1,5 +1,0 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
-
-public enum ReviewTitleType {
-	TITLE, DESCRIPTION
-}
