+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/value/ButtonType.java
@@ -1,0 +1,5 @@
+package com.romanticpipe.reviewcanvas.reviewproperty.domain.value;
+
+public enum ButtonType {
+	NONE, THUMB_UP_WITH_TEXT, THUMB_UP
+}
