+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/value/FilterType.java
@@ -1,0 +1,5 @@
+package com.romanticpipe.reviewcanvas.reviewproperty.domain.value;
+
+public enum FilterType {
+	LIST, DROPDOWN
+}
