+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/value/FocusAreaLayout.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain.value;
 
 public enum FocusAreaLayout {
 	BEST_REVIEW_TOP, BEST_REVIEW_BOTTOM, BEST_REVIEW_LEFT, BEST_REVIEW_RIGHT
