+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/value/ImageReviewAreaLayout.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain.value;
 
 public enum ImageReviewAreaLayout {
 	REVIEW_TOP, REVIEW_BOTTOM, REVIEW_LEFT, REVIEW_RIGHT
