+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/domain/value/SeeMoreButtonType.java
@@ -1,4 +1,4 @@
-package com.romanticpipe.reviewcanvas.reviewproperty.domain;
+package com.romanticpipe.reviewcanvas.reviewproperty.domain.value;
 
 public enum SeeMoreButtonType {
 	SEE_MORE_TOGGLE, SEE_MORE_UNDERLINE, READ_MORE, NO_BUTTON
