+++ domain-module/shopadmin/src/main/java/com/romanticpipe/reviewcanvas/reviewproperty/repository/ReviewTitleRepository.java
@@ -1,11 +1,10 @@
 package com.romanticpipe.reviewcanvas.reviewproperty.repository;
 
-import java.util.Optional;
-
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewTitle;
+import com.romanticpipe.reviewcanvas.reviewproperty.domain.value.ReviewTitleType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewTitle;
-import com.romanticpipe.reviewcanvas.reviewproperty.domain.ReviewTitleType;
+import java.util.Optional;
 
 public interface ReviewTitleRepository extends JpaRepository<ReviewTitle, Integer> {
 	Optional<ReviewTitle> findByShopAdminIdAndReviewTitleType(Integer shopAdminId, ReviewTitleType reviewTitleType);
